+++ example/lib/utils/random.dart
@@ -5,10 +5,21 @@
  */
 
 import 'dart:math';
+import 'dart:ui';
 
 class RandomTool {
   static int genInt({int min = 0, int max = 100}) {
     var x = Random().nextInt(max) + min;
     return x.floor();
+  }
+
+  static Color color() {
+    final random = Random();
+    return Color.fromRGBO(
+      random.nextInt(255),
+      random.nextInt(255),
+      random.nextInt(255),
+      1,
+    );
   }
 }