+++ lib/src/listview/models/listview_observe_model.dart
@@ -12,12 +12,14 @@ import 'listview_observe_displaying_child_model.dart';
 class ListViewObserveModel extends ObserveModel {
   ListViewObserveModel({
     required this.sliverList,
+    required RenderViewportBase viewport,
     required this.firstChild,
     required this.displayingChildModelList,
     required bool visible,
   }) : super(
           visible: visible,
           sliver: sliverList,
+          viewport: viewport,
           innerDisplayingChildModelList: displayingChildModelList,
         );