+++ CHANGELOG.md
@@ -1,5 +1,29 @@
+## 1.16.5
+- ObserverWidget
+  - Improve the processing logic of scroll notification when scrolling with the mouse wheel is not smooth by @qiangjindong in [#48](https://github.com/LinXunFeng/flutter_scrollview_observer/pull/48).
+- Chat Observer
+  - Update `isShrinkWrap` once during initialization by @LinXunFeng. [#47](https://github.com/LinXunFeng/flutter_scrollview_observer/issues/47)
+- ObserverController
+  - Fix unable to jump when sliver is too far away and has no any child by @LinXunFeng. [#45](https://github.com/LinXunFeng/flutter_scrollview_observer/issues/45)
+  - Fix continuous page turning due to incorrect index by @LinXunFeng.
+
+## 1.16.4
+- Supplement missing type conversion adjustments.
+
+## 1.16.3
+- Observation Model
+  - Add viewport property to observation model.
+  - Correct the calculation logic of some values to adapt to the scene of multiple slivers.
+
+## 1.16.2
+- ObserverWidget
+  - Adjust the conversion type of `viewport.offset` to `ScrollPosition` by @LiuDongCai in [#41](https://github.com/LinXunFeng/flutter_scrollview_observer/pull/41).
+- ObserveDisplayingChildModelMixin
+  - Adapt `displayPercentage` to scenes with SliverPersistentHeader by @percival888 in [#43](https://github.com/LinXunFeng/flutter_scrollview_observer/pull/43).
+  - Refine the logic of `calculateDisplayPercentage` method by @LinXunFeng.
+
 ## 1.16.1
-- compatible with flutter 3.13.0
+- Compatible with flutter 3.13.0
 
 ## 1.16.0
 - ObserverController